+++ config/deploy.rb
@@ -1,13 +1,13 @@
-# General settings
+#===[ General settings ]================================================
+
 set :application, "modulesite"
 set :use_sudo, false
 ssh_options[:compression] = false
 default_run_options[:pty] = true
 
-# If you are using Passenger mod_rails uncomment this:
-# if you're still using the script/reapear helper you will need
-# these http://github.com/rails/irs_process_scripts
+#===[ Deploy tasks ]====================================================
 
+# Restart for Phusion Passenger:
 namespace :deploy do
   task :start do ; end
   task :stop do ; end
@@ -17,8 +17,34 @@ namespace :deploy do
 end
 
 desc "Make symlink for database yaml"
+task :symlink_database_yml do
+  run "ln -nfs #{shared_path}/config/database.yml #{release_path}/config/database.yml"
+end
+
+# Callbacks:
 before 'deploy:finalize_update', :symlink_database_yml
 
-task :symlink_database_yml do
-  run "ln -nfs #{shared_path}/config/database.yml #{release_path}/config/database.yml" 
+#===[ Database tasks ]==================================================
+
+namespace :db do
+  desc "Download remote 'production' database and replace existing 'development' database with it."
+  task :use, :roles => :db, :only => {:primary => true} do
+    # NOTE: This assumes that both development and production are using SQLite3 databases.
+    source = "#{user}@#{host}:#{shared_path}/db/production.sqlite3"
+    target = "db/development.sqlite3"
+    backup = target + '.bak'
+
+    sh "cp -a #{target} #{backup}" if File.exist?(target)
+    sh "rsync -uvax #{source} #{target}"
+  end
 end
+
+#===[ Utilities ]=======================================================
+
+# Print the command and then execute it, just like Rake.
+def sh(command)
+  puts command
+  system command unless dry_run
+end
+
+#===[ fin ]=============================================================